AuthService interface created and implemented
